Split differences across version headers
